requirements.txt
streamlit==1.31.0
langchain==0.1.0
langchain-google-genai==0.0.6
langgraph==0.0.20
tavily-python==0.3.0
pandas==2.1.4
python-dotenv==1.0.0